+++ daas/daas_app/api/get_token.py
@@ -1,0 +1,24 @@
+from rest_framework.response import Response
+from rest_framework.authtoken.models import Token
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import AllowAny
+from rest_framework import status
+from django.views.decorators.csrf import csrf_exempt
+from django.contrib.auth import authenticate
+
+
+@csrf_exempt
+@api_view(["POST"])
+@permission_classes((AllowAny,))
+def get_token_view(request):
+    username = request.POST['username']
+    password = request.POST['password']
+    if username is None or password is None:
+        return Response({'message': 'Please provide both username and password'},
+                        status=status.HTTP_400_BAD_REQUEST)
+    user = authenticate(username=username, password=password)
+    if not user:
+        return Response({'message': 'Invalid Credentials'},
+                        status=status.HTTP_404_NOT_FOUND)
+    token, _ = Token.objects.get_or_create(user=user)
+    return Response({'token': token.key}, status=status.HTTP_200_OK)
